Add files via upload
Bubble sort, sometimes referred to as sinking sort, is a simple sorting algorithm that repeatedly steps through the list, compares adjacent elements and swaps them if they are in the wrong order. ... The algorithm, which is a comparison sort, is named for the way smaller or larger elements "bubble" to the top of the list.